developing

+++ pkg/bping/transfer.go
@@ -12,7 +12,7 @@ type MetricValue struct {
 	Endpoint  string      `json:"endpoint"`
 	Metric    string      `json:"metric"`
 	Value     interface{} `json:"value"`
-	Step      int64       `json:"step"`
+	Step      int         `json:"step"`
 	Type      string      `json:"counterType"`
 	Tags      string      `json:"tags"`
 	Timestamp int64       `json:"timestamp"`
